Patch submitted by Jeff Peck <jpjnk4-jdee@yahoo.com>
Fixes jde-parse.el handling of generics and add ability to see full
inner-class name in the mode line:

- I looked a little deeper in JDEE to find why jde-which-method is
failing (failing to recognise any class name that has a Generic
declaration). Turns out Semantic and wisent-java or ok!  but
jde-parse.el does its own regex to find the class name (looking before
the c-parse-state), and the regex does not expect <>!

- For users of nested classes; With this patch you can see the full
inner-class name in the modeline.  [not the FQDN with package name,
but the full outer.inner.inner.inner class name]

Enabled by: (defcustom jde-which-full-class-name ...)
For backward compatibility, the var is initialized to nil.




git-svn-id: https://jdee.svn.sourceforge.net/svnroot/jdee/branches/2.4.0/jde@180 381a1f10-3442-0410-997e-ef331cd04102